Consolidate CLI memory command: eliminate modules/memory layer
Architectural refactor to remove unnecessary indirection between commands and modules directories. The memory.ts command file previously was a thin 16-line shim that delegated all work to modules/memory/index.ts (1360 lines). This created:
- Unclear ownership (business logic split across 2 locations)
- Harder navigation (must jump between files to understand behavior)
- Inconsistent patterns (other commands like watch-status.ts are self-contained)

This change consolidates the full memory command implementation into src/cli/src/commands/memory.ts, making it self-contained and eliminating the src/cli/src/modules/memory/ directory entirely.

Changes:
- Merged src/cli/src/modules/memory/index.ts into src/cli/src/commands/memory.ts
- Removed src/cli/src/modules/memory/ directory
- Updated imports in src/cli/src/cli.ts
- Updated imports in 11 test files
- Updated src/cli/src/modules/index.ts to remove memory namespace export

All tests pass. Build succeeds. Zero functionality change - pure structural improvement.
