+++ src/cli/src/modules/index.ts
@@ -4,7 +4,6 @@ export * from "./plugin-runtime-dependencies.js";
 export * as Feather from "./feather/index.js";
 export * as HotReload from "./hot-reload/index.js";
 export * as Manual from "./manual/index.js";
-export * as Memory from "./memory/index.js";
 export * as Performance from "./performance/index.js";
 export * as QualityReport from "./quality-report/index.js";
 export * as Runtime from "./runtime/index.js";
