+++ src/cli/test/memory-command-format.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMemoryCommand } from "../src/modules/memory/index.js";
+import { createMemoryCommand } from "../src/commands/memory.js";
 import { SuiteOutputFormat } from "../src/cli-core/command-suite-helpers.js";
 import type { ParseOptions } from "commander";
