+++ src/cli/test/memory-command-suite.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMemoryCommand, MemorySuiteName, formatMemorySuiteNameList } from "../src/modules/memory/index.js";
+import { createMemoryCommand, MemorySuiteName, formatMemorySuiteNameList } from "../src/commands/memory.js";
 import type { ParseOptions } from "commander";
 
 const USER_PARSE_OPTIONS: ParseOptions = { from: "user" };
